+++ packages/app/src/components/catalog/EntityPage.tsx
@@ -144,6 +144,7 @@ import {
   EntityNewRelicDashboardCard,
 } from '@backstage/plugin-newrelic-dashboard';
 import { EntityGoCdContent, isGoCdAvailable } from '@backstage/plugin-gocd';
+import { EntityScoreCardContent } from '@oriflame/backstage-plugin-score-card';
 
 import React, { ReactNode, useMemo, useState } from 'react';
 
@@ -701,6 +702,13 @@ const systemPage = (
         </Grid>
         <Grid item md={6}>
           <EntityHasResourcesCard variant="gridItem" />
+        </Grid>
+      </Grid>
+    </EntityLayout.Route>
+    <EntityLayout.Route path="/score" title="Score">
+      <Grid container spacing={3} alignItems="stretch">
+        <Grid item xs={12}>
+          <EntityScoreCardContent />
         </Grid>
       </Grid>
     </EntityLayout.Route>