+++ packages/backend-plugin-api/src/wiring/factories.ts
@@ -44,10 +44,11 @@ export interface BackendPluginConfig<TOptions> {
 
 /** @public */
 export function createBackendPlugin<
-  TOptions extends { [name: string]: unknown } | undefined = undefined,
->(
-  config: BackendPluginConfig<TOptions>,
-): undefined extends TOptions
+  TOptions extends object | undefined = undefined,
+>(config: {
+  id: string;
+  register(reg: BackendRegistrationPoints, options: TOptions): void;
+}): undefined extends TOptions
   ? (options?: TOptions) => BackendFeature
   : (options: TOptions) => BackendFeature {
   return (options?: TOptions) => ({
@@ -68,9 +69,20 @@ export interface BackendModuleConfig<TOptions> {
   ): void;
 }
 
-/** @public */
+/**
+ * @public
+ *
+ * Creates a new backend module for a given plugin.
+ *
+ * The `moduleId` should be equal to the module-specific prefix of the exported name, such
+ * that the full name is `moduleId + PluginId + "Module"`. For example, a GitHub entity
+ * provider module for the `catalog` plugin might have the module ID `'githubEntityProvider'`,
+ * and the full exported name would be `githubEntityProviderCatalogModule`.
+ *
+ * The `pluginId` should exactly match the `id` of the plugin that the module extends.
+ */
 export function createBackendModule<
-  TOptions extends { [name: string]: unknown } | undefined = undefined,
+  TOptions extends object | undefined = undefined,
 >(
   config: BackendModuleConfig<TOptions>,
 ): undefined extends TOptions