+++ plugins/catalog-backend/src/service/CatalogBuilder.ts
@@ -409,7 +409,11 @@ export class CatalogBuilder {
       parser,
       policy,
     });
-    const unauthorizedEntitiesCatalog = new DefaultEntitiesCatalog(dbClient);
+    const stitcher = new Stitcher(dbClient, logger);
+    const unauthorizedEntitiesCatalog = new DefaultEntitiesCatalog(
+      dbClient,
+      stitcher,
+    );
 
     let permissionEvaluator: PermissionEvaluator;
     if ('authorizeConditional' in permissions) {
@@ -453,7 +457,6 @@ export class CatalogBuilder {
       permissions: catalogPermissions,
       rules: this.permissionRules,
     });
-    const stitcher = new Stitcher(dbClient, logger);
 
     const locationStore = new DefaultLocationStore(dbClient);
     const configLocationProvider = new ConfigLocationEntityProvider(config);