+++ plugins/kubernetes-backend/src/types/types.ts
@@ -121,11 +121,22 @@ export interface KubernetesClustersSupplier {
 }
 
 /**
+ * @alpha
+ */
+export interface ServiceLocatorRequestContext {
+  objectTypesToFetch: Set<ObjectToFetch>;
+  customResources: CustomResourceMatcher[];
+}
+
+/**
  * Used to locate which cluster(s) a service is running on
  * @alpha
  */
 export interface KubernetesServiceLocator {
-  getClustersByEntity(entity: Entity): Promise<{ clusters: ClusterDetails[] }>;
+  getClustersByEntity(
+    entity: Entity,
+    requestContext: ServiceLocatorRequestContext,
+  ): Promise<{ clusters: ClusterDetails[] }>;
 }
 
 /**