+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/github/inputProperties.ts
@@ -82,6 +82,11 @@ const allowRebaseMerge = {
   type: 'boolean',
   description: `Allow rebase merges. The default value is 'true'`,
 };
+const allowAutoMerge = {
+  title: 'Allow Auto Merges',
+  type: 'boolean',
+  description: `Allow individual PRs to merge automatically when all merge requirements are met. The default value is 'false'`,
+};
 const collaborators = {
   title: 'Collaborators',
   description: 'Provide additional users or teams with permissions',
@@ -153,6 +158,7 @@ export { access };
 export { allowMergeCommit };
 export { allowRebaseMerge };
 export { allowSquashMerge };
+export { allowAutoMerge };
 export { collaborators };
 export { defaultBranch };
 export { deleteBranchOnMerge };