+++ packages/catalog-client/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/catalog-client
 
+## 1.1.1-next.1
+
+### Patch Changes
+
+- 4f2ac624b4: Renamed argument in `validateEntity` from `location` to `locationRef`
+- Updated dependencies
+  - @backstage/catalog-model@1.1.2-next.1
+  - @backstage/errors@1.1.2-next.1
+
+## 1.1.1-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.2-next.0
+  - @backstage/errors@1.1.2-next.0
+
 ## 1.1.0
 
 ### Minor Changes