+++ plugins/apache-airflow/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-apache-airflow
 
+## 0.2.3-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.11.2-next.1
+  - @backstage/core-plugin-api@1.0.7-next.1
+
+## 0.2.3-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.11.2-next.0
+  - @backstage/core-plugin-api@1.0.7-next.0
+
 ## 0.2.2
 
 ### Patch Changes