+++ plugins/catalog-customized/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @internal/plugin-catalog-customized
 
+## 0.0.3-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.2.0-next.1
+  - @backstage/plugin-catalog@1.6.0-next.1
+
+## 0.0.3-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog@1.5.2-next.0
+  - @backstage/plugin-catalog-react@1.1.5-next.0
+
 ## 0.0.2
 
 ### Patch Changes