+++ plugins/github-issues/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @backstage/plugin-github-issues
 
+## 0.1.2-next.1
+
+### Patch Changes
+
+- 8c7bff2bb4: Updated the `luxon` dependency to 3.x
+- 719ccbb963: Properly filter on relations instead of the spec, when finding by owner
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.2.0-next.1
+  - @backstage/core-components@0.11.2-next.1
+  - @backstage/core-plugin-api@1.0.7-next.1
+  - @backstage/catalog-model@1.1.2-next.1
+  - @backstage/errors@1.1.2-next.1
+  - @backstage/integration@1.3.2-next.1
+  - @backstage/theme@0.2.16
+
+## 0.1.2-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.2-next.0
+  - @backstage/core-components@0.11.2-next.0
+  - @backstage/plugin-catalog-react@1.1.5-next.0
+  - @backstage/core-plugin-api@1.0.7-next.0
+  - @backstage/errors@1.1.2-next.0
+  - @backstage/integration@1.3.2-next.0
+  - @backstage/theme@0.2.16
+
 ## 0.1.1
 
 ### Patch Changes