+++ plugins/periskop-backend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-periskop-backend
 
+## 0.1.8-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.15.2-next.1
+  - @backstage/config@1.0.3-next.1
+
+## 0.1.8-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.15.2-next.0
+  - @backstage/config@1.0.3-next.0
+
 ## 0.1.7
 
 ### Patch Changes