+++ plugins/playlist-common/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-playlist-common
 
+## 0.1.1-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-permission-common@0.6.5-next.1
+
+## 0.1.1-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-permission-common@0.6.5-next.0
+
 ## 0.1.0
 
 ### Minor Changes